# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/weather/current": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Weather"
                ],
                "summary": "Getting Weather Info for now",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Location (ex. 'Moscow')",
                        "name": "location",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/weather/daily": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Weather"
                ],
                "summary": "Getting Daily Weather Info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Location (ex. 'Moscow')",
                        "name": "location",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Date (ex. '2000-05-30')",
                        "name": "date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/weather/hourly": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Weather"
                ],
                "summary": "Getting Hourly Weather Info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Location (ex. 'Moscow')",
                        "name": "location",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Date (ex. '2000-05-30')",
                        "name": "date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}